+++ doc/changelog.md
@@ -14,6 +14,7 @@ Jump to:
 Description
 
 - RequestBatch rewrite
+- Fix regression on hostlist param to DragonRunRequest
 - Fix dragon build logging bug
 - Merge core refactor into MLI feature branch
 - Implement asynchronous notifications for shared data